move to public header
