+++ include/tmc/detail/ex_manual_st.ipp
@@ -6,10 +6,10 @@
 #pragma once
 
 #include "tmc/detail/impl.hpp" // IWYU pragma: keep
-#include "tmc/detail/qu_mpsc.hpp"
 #include "tmc/detail/thread_locals.hpp"
 #include "tmc/ex_any.hpp"
 #include "tmc/ex_manual_st.hpp"
+#include "tmc/qu_unbounded_mpsc.hpp"
 #include "tmc/work_item.hpp"
 
 #include <cassert>
