+++ include/tmc/ex_braid.hpp
@@ -9,8 +9,8 @@
 #include "tmc/aw_resume_on.hpp"
 #include "tmc/detail/compat.hpp"
 #include "tmc/detail/concepts_awaitable.hpp"
-#include "tmc/detail/qu_mpsc.hpp"
 #include "tmc/ex_any.hpp"
+#include "tmc/qu_unbounded_mpsc.hpp"
 #include "tmc/task.hpp"
 #include "tmc/utils.hpp"
 #include "tmc/work_item.hpp"
